Migrate VirtualScroll usage to masonry's new usize-indexed API
Upstream masonry (branch=main) replaced VirtualScroll's i64-anchor-domain
API with a usize-indexed, fixed-len one, and renamed border_box_size()/
content_box_size() to border_box()/content_box() (now returning Rect
instead of Size). Bumps the pinned masonry commit accordingly and updates
every call site:

- collection/{ids,scroll,body,body_view}.rs: virtual_scroll(Range<i64>, ..)
  -> virtual_scroll(usize, ..), scroll_idx_to_slice removed (the callback
  index is now already the slice position), overwrite_anchor -> scroll_to,
  VirtualScrollAction is now an enum (Fetch/Scroll) with accessor methods
  instead of public fields.
- ~30 widget files: border_box_size()/content_box_size() -> .size() on the
  new Rect-returning border_box()/content_box().

Known regression tracked separately: masonry's VirtualScroll now handles
ArrowUp/Down itself for built-in scroll-by-arrow-key, which intercepts the
key event before CollectionBodyWidget's row-focus navigation runs. 5 tests
in collection::body::tests are left failing (not weakened) to keep
flagging it — see #175.

+++ src/collection/scroll.rs
@@ -62,20 +62,20 @@ impl ScrollState {
 }
 
 /// Clamps a requested scroll index to `0..row_count`, converted to the
-/// `i64` anchor domain masonry's `VirtualScroll` uses (saturating, same
+/// `usize` domain masonry's `VirtualScroll` uses (saturating, same
 /// rationale as `scroll_range_end` in `view.rs`). `None` when the grid
 /// is empty — there is no row to anchor to.
-pub(crate) fn clamp_scroll_index(index: u64, row_count: u64) -> Option<i64> {
+pub(crate) fn clamp_scroll_index(index: u64, row_count: u64) -> Option<usize> {
     if row_count == 0 {
         return None;
     }
     let clamped = index.min(row_count - 1);
-    Some(i64::try_from(clamped).unwrap_or(i64::MAX))
+    Some(usize::try_from(clamped).unwrap_or(usize::MAX))
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{ScrollState, clamp_scroll_index};
+    use super::{clamp_scroll_index, ScrollState};
 
     #[test]
     fn default_state_has_no_pending_request() {
